refactor: move deps to core folder
