+++ app/api/core/infrastructure/factories/DeleteTemplateUseCaseFactory.ts
@@ -2,7 +2,7 @@ import { DefaultTransactionManager } from 'api/common.v2/database/data_source_de
 import { DefaultTemplatesDataSource } from 'api/templates.v2/database/data_source_defaults';
 import { DefaultSettingsDataSource } from 'api/settings.v2/database/data_source_defaults';
 import { DeleteTemplateUseCase } from 'api/core/application/DeleteTemplate';
-import { applicationEventsBus } from 'api/eventsbus';
+import { applicationEventsBus } from 'api/core/libs/eventsbus';
 import { DefaultEntitiesDataSource } from 'api/entities.v2/database/data_source_defaults';
 import { DefaultTranslationsDataSource } from 'api/i18n.v2/database/data_source_defaults';
 import { permissionsContext } from 'api/permissions/permissionsContext';
