+++ app/api/paragraphExtraction/infrastructure/PXEntityDeletedListener.ts
@@ -1,4 +1,4 @@
-import { EventsBus } from 'api/eventsbus';
+import { EventsBus } from 'api/core/libs/eventsbus';
 import { getConnection } from 'api/common.v2/database/getConnectionForCurrentTenant';
 import { DefaultTransactionManager } from 'api/common.v2/database/data_source_defaults';
 import { EntityDeletedEvent } from 'api/entities/events/EntityDeletedEvent';
