+++ app/api/services/pdfsegmentation/eventListeners.ts
@@ -1,4 +1,4 @@
-import { EventsBus } from 'api/eventsbus';
+import { EventsBus } from 'api/core/libs/eventsbus';
 import { FilesDeletedEvent } from 'api/files/events/FilesDeletedEvent';
 import { SegmentationModel } from './segmentationModel';
