+++ app/api/suggestions/listeners/afterEntityUpdatedListener.ts
@@ -1,4 +1,4 @@
-import { EventsBus } from 'api/eventsbus';
+import { EventsBus } from 'api/core/libs/eventsbus';
 import { SettingsDataSource } from 'api/settings.v2/contracts/SettingsDataSource';
 import { Logger } from 'api/log.v2/contracts/Logger';
 import { EntityUpdatedEvent } from 'api/entities/events/EntityUpdatedEvent';
