+++ app/api/suggestions/specs/afterFileCreatedListener.spec.ts
@@ -1,4 +1,4 @@
-import { EventsBus } from 'api/eventsbus';
+import { EventsBus } from 'api/core/libs/eventsbus';
 import { FileUpdatedEvent } from 'api/files/events/FileUpdatedEvent';
 import { getFixturesFactory } from 'api/utils/fixturesFactory';
 import { TestUtils } from 'api/common.v2/utils/Test';
